requirements.txt
flask
line-bot-sdk
python-dotenv
torch
transformers
sentencepiece
faiss-cpu
numpy
tqdm